Users, share links and add button

+++ blog/views.py
@@ -1,7 +1,10 @@
+from django.http import Http404
 from django.http import HttpResponse
 from django.http import HttpResponseRedirect
 from django.shortcuts import render, get_object_or_404,redirect
 from django.contrib import messages
+
+from urllib.parse import quote_plus
 
 
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
@@ -15,9 +18,14 @@ from .forms import PostForm
 
 
 def post_create(request):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    if not request.user.is_authenticated():
+        raise Http404
     form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         instance = form.save(commit=False)
+        instance.user = request.user
         instance.save()
         messages.success(request, "Successfully created")
         return HttpResponseRedirect(instance.get_absolute_path())
@@ -29,8 +37,8 @@ def post_create(request):
     }
     return render(request, "post_form.html", context)
 
-def post_detail(request, id):
-    instance = get_object_or_404(Post, id=id)
+def post_detail(request, slug):
+    instance = get_object_or_404(Post, slug=slug)
     context = {
         "title": instance.title,
         "instance": instance,
@@ -59,8 +67,10 @@ def post_list(request):
     return render(request, "post_list.html", context)
 
 
-def post_update(request, id=None):
-    instance = get_object_or_404(Post, id=id)
+def post_update(request, slug=None):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    instance = get_object_or_404(Post, slug=slug)
     form = PostForm(request.POST or None, request.FILES or None, instance=instance)
     if form.is_valid():
         instance = form.save(commit=False)
@@ -74,8 +84,10 @@ def post_update(request, id=None):
     }
     return render(request, "post_form.html", context)
 
-def post_delete(request, id=None):
-    instance = get_object_or_404(Post, id=id)
+def post_delete(request, slug=None):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    instance = get_object_or_404(Post, slug=slug)
     instance.delete()
     messages.success(request, "Successfully deleted")
     return redirect("posts:list")
